two triangle
